+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.2] - 2025-11-17
+
+### Added
+
+- Text runs for table cell for more accurate formatting
+- Deduplication for iframe parser
+
+### Changed
+
+- `COORDINATE_BUFFER` from `1.02` to `1.03` bc too small
+
+## [0.2.1] - 2025-11-17
+
+### Changed
+
+- Bug fixes i guess
 
 ## [0.2.0] - 2025-11-17